Don't hide JVM symbols by default on FreeBSD
The symbols are needed for producing accurate stack traces for native
code. The way FreeBSD loads dynamic libs, and looks up symbols, it is
not able to find symbols that are marked as hidden (i.e. local.)

This patch changes the configuration for FreeBSD, so that the JVM will
be built with default symbol visibility.

This work was sponsored by: The FreeBSD Foundation