Elf backtrace code fails in debug mode if debug info missing
The function DwarfFile::DebugArranges::find_compilation_unit_offset will
return false if it's not able to locate the debug information for a
symbol. This can happen if a module is built without debug symbols, as
is typical for OS provided shared libraries.

This is fine, and the call site for this function handles the failure
just fine.

In debug builds however, tha function will assert that it was able to
find the debug info. This cause the StackWalkNativeToJava test to fail
on FreeBSD, because it's not able to find debug information for
pthread_create in libthr.so.

While it's possible to solve this by installing debug info for the OS, I
don't think that should be a requirement for the tests to pass.

This patch simply replaces the assert with a DWARF_LOG_INFO construct
instead.

This work was sponsored by: The FreeBSD Foundation