Sync SA class sharing workaround after JDK-8320935
In commit 4c96aac9c0aa (JDK-8320935) the symbol to locate the shared
archive path was changes. This is used by the common ps core routines
to find the reference to libjvm in a core dump.

This patch updates the BSD implementation to use the correct symbol by
useing the same code as for Linux.

This work was sponsored by: The FreeBSD Foundation