Improve FreeBSD impl of getCommittedVirtualMemorySize
Use getrlimiusage instead of relying on procfs to get the size of
committed memory. This should improve performance, as well as reduce our
reliance on procfs.

The implementation was also moved from the generic unix/native to the
bsd/native source tree.

This work was sponsored by: The FreeBSD Foundation