Reset OpenJDK version info to 25.0.1
The bsd port version tagged 25.0.1+8-1 is contains code from 25.0.2, and
identifies itself internally as such with a release date of January 10
next year.

This patch _only_ resets the version number, and sets todays date as
release date. It still contains the bug-fixes that currently is
scheduled for 25.0.2.

This work was sponsored by: The FreeBSD Foundation